feat: cave objects (#107)
* feat: cave object generation and bufixes and refactors

* feat: sound addon

* feat: fence collision

* yes

* refactor: storngertyping

+++ scenes/player/player.gd
@@ -219,8 +219,9 @@ func damage_victim(victim, damage):
 	if victim is Destroyable:
 		if victim.required_tool == hotbar.get_held_item() or victim.required_tool == null:
 			if victim.take_damage(damage):
-				var tile_pos = $"../ObjectLayer".local_to_map(victim.global_position)
-				get_parent().delete_object_at(tile_pos)
+				var tile_pos = victim.get_parent().local_to_map(victim.global_position)
+				if get_parent().has_method(&"delete_object_at"):
+					get_parent().delete_object_at(tile_pos)
 				inventory.add_item(victim.get_drop(), 1)
 	
 func consume(item: InventoryItem, amount: int) -> void:
